Ensure `tsconfig.json` is referenced for linting
When `tsc` is called with a file name, it ignores settings in
`tsconfig.json`. This commit defines the files to be checked within the
config file itself to address this issue. Our call to `tslint` is now
also adjusted to read the config file for the list of files to lint.