Update npm dependencies
For now we're not updating `knockout` or `@types/knockout`.  As of
`knockout@3.5.0`, the package bundles its own TypeScript typings and we
won't need `@types/knockout` anymore (yay!), *but* we'll need to adjust
our app code to reference the bundled typings when making that upgrade.
Those changes will get their own commit.