Fix vulnerability by updating tar (subdependency)
The `tar` package had a vulnerability prior to version `4.4.2`. Tar is a
dependency of `node-sass`, which we directly depend on, but it hasn't
been updated to depend on the fixed version yet.

Fixing this required manually editing `package-lock.json`, deleting
`node_modules/`, and reinstalling using `npm install`.

The vulnerability advisory: https://www.npmjs.com/advisories/803

Instructions referenced for manually fixing this:
* https://github.com/sass/node-sass/issues/2625#issuecomment-482224111
* https://github.com/sass/node-sass/issues/2625#issuecomment-482519446